chore: rename seeder files
